End-To-End Projects/LLM Apps With Chainlit/requirements.txt
chainlit
groq
python-dotenv
-e .